lex-experiments/requirements.txt
boto3
langchain-community
langchain-core
# langchain
streamlit
audio-recorder-streamlit
langchain-aws
moviepy
pillow
